Refactor authentication module to include BasicAuth class
This commit adds the BasicAuth class to the authentication module in order to support basic authentication. The BasicAuth class includes methods for extracting and decoding the base64 authorization header, as well as extracting user credentials and retrieving the corresponding user object. This change aligns with recent user commits that focused on improving the authentication logic and readability of the code.

This commit message follows the established convention of using a prefix to indicate the type of change, in this case, "refactor" for code refactoring.